conversion-system/python-system/requirements.txt
flask>=3.0.0
httpx>=0.27.0
twilio>=9.0.0
python-dotenv>=1.0.0
imapclient>=3.0.0
mcp>=0.1.0
user-agents>=2.2.0
psutil>=5.9.0
websockets>=12.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pipecat-ai>=0.0.40
openai>=1.12.0
deepgram-sdk>=3.2.7
numpy>=1.26.3
scipy>=1.12.0
loguru>=0.7.2
python-dateutil>=2.8.2
python-multipart>=0.0.6
aiohttp>=3.9.3
pyngrok>=7.1.0
